tietokanta poistettu ja luotu uudelleen. toimittajan haku, luonti ja poisto toiminto luotu ja toimii.

+++ Kurssi/app/views.py
@@ -48,33 +48,48 @@ def tuotelistaview(request):
 
 
 def lisäätuote(request):
-    a = request.POST['tuotenimi']
-    b = request.POST['painoperkappale']
-    c = request.POST['kappalehinta']
-    d = request.POST['tuotteitavarastossa']
-    e = request.POST['toimittaja']
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        a = request.POST['tuotenimi']
+        b = request.POST['painoperkappale']
+        c = request.POST['kappalehinta']
+        d = request.POST['tuotteitavarastossa']
+        e = request.POST['toimittaja']
     
-    Tuote(tuotenimi = a, painoperkappale = b, kappalehinta = c, tuotteitavarastossa = d, toimittaja = Toimittaja.objects.get(id = e)).save()
-    return redirect(request.META['HTTP_REFERER'])
+        Tuote(tuotenimi = a, painoperkappale = b, kappalehinta = c, tuotteitavarastossa = d, toimittaja = Toimittaja.objects.get(id = e)).save()
+        return redirect(request.META['HTTP_REFERER'])
 
 def vahvistatuotepoisto(request, id):
-    tuote = Tuote.objects.get(id = id)
-    context = {'tuote': tuote}
-    return render (request,"tuotepoisto.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        tuote = Tuote.objects.get(id = id)
+        context = {'tuote': tuote}
+        return render (request,"tuotepoisto.html",context)
 
 
 def tuotepoisto(request, id):
-    Tuote.objects.get(id = id).delete()
-    return redirect(tuotelistaview)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        Tuote.objects.get(id = id).delete()
+        return redirect(tuotelistaview)
 
 
 def edit_tuote_get(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
         tuote = Tuote.objects.get(id = id)
         context = {'tuote': tuote}
         return render (request,"edit_tuote.html",context)
 
 
 def edit_tuote_post(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
         item = Tuote.objects.get(id = id)
         item.kappalehinta = request.POST['kappalehinta']
         item.tuotteitavarastossa = request.POST['tuotteitavarastossa']
@@ -83,10 +98,13 @@ def edit_tuote_post(request, id):
 
 
 def tuotteet_filtered(request, id):
-    tuotelista = Tuote.objects.all()
-    filteredtuotteet = tuotelista.filter(toimittaja = id)
-    context = {'tuotteet': tuotteet_filtered}
-    return render (request,"tuotelista.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        tuotelista = Tuote.objects.all()
+        filteredtuotteet = tuotelista.filter(toimittaja = id)
+        context = {'tuotteet': tuotteet_filtered}
+        return render (request,"tuotelista.html",context)
 
 
 
@@ -101,18 +119,45 @@ def toimittajalistaview(request):
 
 
 def lisäätoimittaja(request):
-    a = request.POST['yritysnimi']
-    b = request.POST['yhteyshenkiö']
-    c = request.POST['osoite']
-    d = request.POST['puhelin']
-    e = request.POST['sähköposti']
-    f = request.POST['maa']
-    Toimittaja(yritysnimi = a, yhteyshenkiö = b, osoite = c, puhelin = d, sähköposti = e, maa = f).save()
-    return redirect(request.META['HTTP_REFERER'])
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        a = request.POST['yritysnimi']
+        b = request.POST['yhteyshenkilö']
+        c = request.POST['osoite']
+        d = request.POST['puhelin']
+        e = request.POST['sähköposti']
+        f = request.POST['maa']
+        Toimittaja(yritysnimi = a, yhteyshenkilö = b, osoite = c, puhelin = d, sähköposti = e, maa = f).save()
+        return redirect(request.META['HTTP_REFERER'])
 
 
 def etsitoimittaja(request):
-    search = request.POST['search']
-    filtered = Toimittaja.objects.filter(yritysnimi__icontains=search)
-    context = {'toimittajat': filtered}
-    return render (request,"toimittajalista.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        search = request.POST['search']
+        filtered = Toimittaja.objects.filter(yritysnimi__icontains=search)
+        context = {'toimittajat': filtered}
+        return render (request,"toimittajalista.html",context)
+    
+def vahvistatoimittajapoisto(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        toimittaja = Toimittaja.objects.get(id = id)
+        context = {'toimittaja': toimittaja}
+        return render (request,"toimittajapoisto.html",context)
+
+
+def toimittajapoisto(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        try:
+            toimittaja = Toimittaja.objects.get(id=id)  # Hae toimittaja id:n mukaan
+            toimittaja.delete()  # Poista toimittaja
+            return redirect(toimittajalistaview)  # Ohjaa takaisin toimittajalistaan
+        except Toimittaja.DoesNotExist:
+            # Jos toimittajaa ei löydy, ohjataan virhesivulle tai näyttää viestin
+            return render(request, 'error.html', {'message': 'Toimittajaa ei löytynyt!'})
